+++ OSSMediatorCollector/pkg/ndacapis/um_apis_test.go
@@ -17,7 +17,7 @@ import (
 	"testing"
 	"time"
 
-	jwt "github.com/dgrijalva/jwt-go"
+	jwt "github.com/golang-jwt/jwt/v4"
 	log "github.com/sirupsen/logrus"
 )