+++ OSSMediatorCollector/pkg/utils/file_utils.go
@@ -10,7 +10,6 @@ import (
 	"collector/pkg/config"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path"
 	"sort"
@@ -24,14 +23,12 @@ import (
 var (
 	//CurrentTime
 	CurrentTime = time.Now
-	//mux         sync.Mutex
 )
 
 const (
 	//File extension for writing response
 	fileExtension = ".json"
 
-	//
 	fmdataResponseType = "fmdata"
 	pmdataResponseType = "pmdata"
 	nhgResponseType    = "network-hardware-groups"
@@ -73,18 +70,24 @@ func writeFile(fileName string, data []byte) error {
 //CreateResponseDirectory creates directory named path, along with any necessary parents.
 // If the directory creation fails it will terminate the program.
 func CreateResponseDirectory(basePath string, api string) {
-	path := basePath + "/" + path.Base(api)
-	log.Infof("Creating %s directory", path)
-	err := os.MkdirAll(path, os.ModePerm)
+	dirPath := basePath + "/" + path.Base(api)
+	log.Infof("Creating %s directory", dirPath)
+	err := os.MkdirAll(dirPath, os.ModePerm)
 	if err != nil {
-		log.WithFields(log.Fields{"error": err}).Fatalf("Error while creating %s", path)
+		log.WithFields(log.Fields{"error": err}).Fatalf("Error while creating %s", dirPath)
 	}
 }
 
 //WriteResponse writes the data in json format to responseDest directory.
 func WriteResponse(user *config.User, api *config.APIConf, data interface{}, id string, txnID uint64) error {
-	//mux.Lock()
-	//defer mux.Unlock()
+	//formattedData, err := json.MarshalIndent(data, "", "  ")
+	//if err != nil {
+	//	log.WithFields(log.Fields{"tid": txnID, "error": err}).Errorf("Unable to indent received data, error: %v", err)
+	//	return err
+	//}
+	//data = nil
+
+	log.WithFields(log.Fields{"tid": txnID}).Info("starting response write")
 	fileName := path.Base(api.API)
 	if fileName == fmdataResponseType || fileName == pmdataResponseType {
 		if api.MetricType != "" {
@@ -117,25 +120,28 @@ func WriteResponse(user *config.User, api *config.APIConf, data interface{}, id 
 	}
 	fileName = name + fileExtension
 
-	formattedData, err := json.MarshalIndent(data, "", "  ")
-	if err != nil {
-		log.WithFields(log.Fields{"tid": txnID, "error": err}).Errorf("Unable to indent received data, error: %v", err)
-		return err
-	}
-
 	log.WithFields(log.Fields{"tid": txnID}).Infof("Writing response to file %s for %s", fileName, user.Email)
-	err = writeFile(fileName, formattedData)
+	file, _ := os.OpenFile(fileName, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0666)
+	defer file.Close()
+	encoder := json.NewEncoder(file)
+	err := encoder.Encode(data)
 	if err != nil {
 		log.WithFields(log.Fields{"tid": txnID, "error": err}).Errorf("Writing response to file %s for %s failed", fileName, user.Email)
 		return err
 	}
+
+	//err = writeFile(fileName, formattedData)
+	//if err != nil {
+	//	log.WithFields(log.Fields{"tid": txnID, "error": err}).Errorf("Writing response to file %s for %s failed", fileName, user.Email)
+	//	return err
+	//}
 	return nil
 }
 
 //retrieves the last received metric time from file per API
 func getLastReceivedDataTime(user *config.User, api *config.APIConf, nhgID string) string {
 	//Reading start time value from file
-	fileName := path.Base(api.API)
+	fileName := "checkpoints/" + path.Base(api.API)
 	if api.MetricType != "" {
 		fileName = fileName + "_" + api.MetricType
 	}
@@ -146,7 +152,7 @@ func getLastReceivedDataTime(user *config.User, api *config.APIConf, nhgID strin
 	if nhgID != "" {
 		fileName = fileName + "_" + nhgID
 	}
-	data, err := ioutil.ReadFile(fileName)
+	data, err := os.ReadFile(fileName)
 	if err != nil || len(data) == 0 {
 		return ""
 	}
@@ -175,6 +181,7 @@ func StoreLastReceivedDataTime(user *config.User, data interface{}, api *config.
 			receivedData[metricTime] = struct{}{}
 		}
 	}
+	data = nil
 	var eventTimes []time.Time
 	for key := range receivedData {
 		eventTime, err := time.Parse(eventTimeFormatBaiCell, key)
@@ -197,7 +204,7 @@ func StoreLastReceivedDataTime(user *config.User, data interface{}, api *config.
 	sort.Slice(eventTimes, func(i, j int) bool { return eventTimes[i].Before(eventTimes[j]) })
 	latestTime := truncateSeconds(eventTimes[len(eventTimes)-1]).Format(time.RFC3339)
 
-	fileName := path.Base(api.API)
+	fileName := "checkpoints/" + path.Base(api.API)
 	if api.MetricType != "" {
 		fileName = fileName + "_" + api.MetricType
 	}